Removed old agni interface
