+++ utils/agni.py
@@ -5,28 +5,18 @@ from utils.helper import *
 from utils.constants import *
 from utils.logs import GetLogfilePath, GetCurrentLogfileIndex
 
-import tomlkit as toml
 from juliacall import Main as jl
 
 log = logging.getLogger("PROTEUS")
 
-def _try_agni(loops_total:int, dirs:dict, COUPLER_options:dict, 
-              hf_row:dict, make_plots:bool, easy_start:bool,
-              linesearch:bool, dx_max:float, resume_prev:bool)->bool:
 
-    # ---------------------------
-    # Setup values to be provided to AGNI
-    # ---------------------------
-    agni_debug = bool(log.getEffectiveLevel() == logging.DEBUG)
-    try_spfile = os.path.join(dirs["output"] , "runtime.sf")
-    
-    # Get stellar spectrum at TOA
-    sflux_files = glob.glob(dirs["output"]+"/data/*.sflux")
-    sflux_times = [ int(s.split("/")[-1].split(".")[0]) for s in sflux_files]
-    sflux_tlast = sorted(sflux_times)[-1]
-    sflux_path  = dirs["output"]+"/data/%d.sflux"%sflux_tlast
+def ActivateEnv(agni_dir:str):
+    jl.seval("using Pkg")
+    jl.Pkg.activate(agni_dir)
+    jl.seval("using AGNI")
 
-    # store VMRs
+
+def ConstructVolDict(hf_row:dict, COUPLER_options:dict):
     vol_dict = {}
     for vol in volatile_species:
         if COUPLER_options[vol+"_included"]:
@@ -38,55 +28,67 @@ def _try_agni(loops_total:int, dirs:dict, COUPLER_options:dict,
         UpdateStatusfile(dirs, 20)
         raise Exception("All volatiles have a volume mixing ratio of zero")
     
-    if COUPLER_options["tropopause"] not in [0,1]:
-        UpdateStatusfile(dirs, 20)
-        raise Exception("Tropopause type not supported by AGNI")
+    return vol_dict
 
-    # ---------------------------
-    # Make configuration file
-    # ---------------------------
 
-    log.debug("Write cfg file")
+def InitAtmos(dirs:dict, COUPLER_options:dict, hf_row:dict):
+    """Initialise atmosphere struct for use by AGNI.
+    
+    Does not set the temperature profile.
 
-    # Read base AGNI configuration file
-    cfg_base = os.path.join(dirs["utils"] , "templates", "init_agni.toml")
-    with open(cfg_base, 'r') as hdl:
-        cfg_toml = toml.load(hdl) 
+    Parameters
+    ----------
+        dirs : dict
+            Dictionary containing paths to directories
+        COUPLER_options : dict
+            Configuration options and other variables
+        hf_row : dict
+            Dictionary containing simulation variables for current iteration
 
-    # Setup new AGNI configuration file for this run
-    cfg_this = os.path.join(dirs["output"] , "agni_recent.toml")
+    Returns
+    ----------
+        atmos : atmosphere.Atmos_t
+            Atmosphere struct 
 
-    # Set title 
-    cfg_toml["title"] = "PROTEUS runtime step %d"%loops_total
+    """
 
-    # Set planet 
-    cfg_toml["planet"]["tmp_surf"] =        hf_row["T_surf"]
-    cfg_toml["planet"]["instellation"] =    hf_row["F_ins"]
-    cfg_toml["planet"]["s0_fact"] =         COUPLER_options["asf_scalefactor"]
-    cfg_toml["planet"]["albedo_b"] =        COUPLER_options["albedo_pl"]
-    cfg_toml["planet"]["zenith_angle"] =    COUPLER_options["zenith_angle"]
-    cfg_toml["planet"]["albedo_s"] =        COUPLER_options["albedo_s"]
-    cfg_toml["planet"]["gravity"] =         hf_row["gravity"]
-    cfg_toml["planet"]["radius"] =          hf_row["R_planet"]
 
-    # set composition
-    cfg_toml["composition"]["p_surf"] =     hf_row["P_surf"]
-    cfg_toml["composition"]["p_top"] =      COUPLER_options["P_top"]
-    cfg_toml["composition"]["vmr_dict"] =   vol_dict
+    # Create atmos struct 
+    atmos = jl.AGNI.atmosphere.Atmos_t()
 
+    spfile_name = "AGNI/res/spectral_files/Dayspring/48/Dayspring.sf"
+    star_file = "AGNI/res/stellar_spectra/sun.txt"
+
+    # Stellar spectrum path
+    sflux_files = glob.glob(dirs["output"]+"/data/*.sflux")
+    sflux_times = [ int(s.split("/")[-1].split(".")[0]) for s in sflux_files]
+    sflux_path  = dirs["output"]+"/data/%d.sflux"%int(sorted(sflux_times)[-1])
+
+    # Spectral file path
+    try_spfile = os.path.join(dirs["output"] , "runtime.sf")
+    if os.path.exists(try_spfile):
+        # exists => don't modify it
+        input_sf =      try_spfile
+        input_star =    ""   
+    else:
+        # doesn't exist => AGNI will copy it + modify as required
+        input_sf =      os.path.join(dirs["fwl"], COUPLER_options["spectral_file"])
+        input_star =    sflux_path
+
+   
+    # Chemistry 
     chem_type = COUPLER_options["atmosphere_chemistry"]
-    if chem_type > 0:
-        # any chemistry
-        cfg_toml["plots"]["mixing_ratios"] = True
+    include_all = False
+    if chem_type == 1:
+        # equilibrium
+        include_all = True
 
-        if chem_type == 1:
-            # equilibrium
-            cfg_toml["composition"]["chemistry"]   = chem_type
-            cfg_toml["composition"]["include_all"] = True
-
-        elif chem_type >= 2:
-            # kinetics 
-            raise Exception("Chemistry type %d unsupported by AGNI"%chem_type)
+    elif chem_type >= 2:
+        # kinetics 
+        raise Exception("Chemistry type %d unsupported by AGNI"%chem_type)
+    
+    # composition
+    vol_dict = ConstructVolDict(hf_row, COUPLER_options)
     
     # set condensation
     condensates = []
@@ -106,105 +108,89 @@ def _try_agni(loops_total:int, dirs:dict, COUPLER_options:dict,
             if k == gas_min:
                 continue 
             condensates.append(k)
-    cfg_toml["composition"]["condensates"] = condensates
 
-    if len(condensates) > 0:
-        cfg_toml["plots"]["mixing_ratios"] = make_plots
+    # Setup struct 
+    return_success = jl.AGNI.atmosphere.setup_b(atmos, 
+                                                
+                        dirs["agni"], dirs["output"], input_sf,
 
-    # Set files
-    cfg_toml["files"]["output_dir"] =       os.path.join(dirs["output"])
-    if os.path.exists(try_spfile):
-        # exists => don't modify it
-        cfg_toml["files"]["input_sf"] =     try_spfile
-        cfg_toml["files"]["input_star"] =   ""   
-    else:
-        # doesn't exist => AGNI will copy it + modify as required
-        cfg_toml["files"]["input_sf"] =     os.path.join(dirs["fwl"],
-                                                         COUPLER_options["spectral_file"])
-        cfg_toml["files"]["input_star"] =   sflux_path
+                        hf_row["F_ins"], 
+                        COUPLER_options["asf_scalefactor"], 
+                        COUPLER_options["albedo_pl"], 
+                        COUPLER_options["zenith_angle"],
+
+                        hf_row["T_surf"], 
+                        hf_row["gravity"], hf_row["R_planet"],
+                        
+                        int(COUPLER_options["atmosphere_nlev"]), 
+                        hf_row["P_surf"], 
+                        COUPLER_options["P_top"],
+
+                        vol_dict, "",
+
+                        flag_rayleigh=bool(COUPLER_options["rayleigh"] == 1),
+                        flag_cloud=bool(COUPLER_options["water_cloud"] == 1),
+                        
+                        albedo_s=COUPLER_options["albedo_s"],
+                        condensates=condensates,
+                        use_all_gases=include_all,
+
+                        skin_d=COUPLER_options["skin_d"], skin_k=COUPLER_options["skin_k"],
+                        tmp_magma=hf_row["T_surf"]
+                        )
+
+    # Allocate arrays 
+    jl.AGNI.atmosphere.allocate_b(atmos,star_file)
+
+    return atmos
+                                    
+
+
+def UpdateProfile(atmos, hf_row:dict, COUPLER_options:dict, resume:bool):
+    """Update atmosphere struct.
     
-    # Set execution
-    cfg_toml["execution"]["num_levels"] =   COUPLER_options["atmosphere_nlev"]
-    cfg_toml["execution"]["rayleigh"] =     bool(COUPLER_options["rayleigh"] == 1)
-    cfg_toml["execution"]["cloud"] =        bool(COUPLER_options["water_cloud"] == 1)
-    cfg_toml["execution"]["linesearch"] =   linesearch
-    cfg_toml["execution"]["easy_start"] =   easy_start
-    cfg_toml["execution"]["dx_max"] =       dx_max
+    Sets the new surface boundary conditions and composition.
 
-    if (loops_total > 1) and resume_prev:
-        # If solving for RCE and are current inside the init stage, use old T(p)
-        # as initial guess for solver.
-        ncdfs = glob.glob(os.path.join(dirs["output"], "data","*_atm.nc"))
-        ncdf_times = [float(f.split("/")[-1].split("_")[0]) for f in ncdfs]
-        nc_path = ncdfs[np.argmax(ncdf_times)]
+    Parameters
+    ----------
+        atmos : atmosphere.Atmos_t
+            Atmosphere struct 
+        COUPLER_options : dict
+            Configuration options and other variables
+        hf_row : dict
+            Dictionary containing simulation variables for current iteration
+        resume : bool
+            Resume from previous temperature profile 
 
-        log.debug("Initialise from last T(p)")
-        cfg_toml["execution"]["initial_state"] = ["ncdf", nc_path]
+    Returns
+    ----------
+        atmos : atmosphere.Atmos_t
+            Atmosphere struct 
 
-    else:
-        log.debug("Initialise isothermal")
-        cfg_toml["execution"]["initial_state"] = ["iso", "%.2f"%(hf_row["T_surf"]-1.0)]
-        
-    # Solution stuff 
-    surf_state = int(COUPLER_options["atmosphere_surf_state"])
-    if not (0 <= surf_state <= 3):
-        UpdateStatusfile(dirs, 20)
-        raise Exception("Invalid surface state %d" % surf_state)
+    """
 
-    # CBL case
-    if surf_state == 2:
-        cfg_toml["planet"]["skin_k"] =    COUPLER_options["skin_k"]
-        cfg_toml["planet"]["skin_d"] =    COUPLER_options["skin_d"]
-        cfg_toml["planet"]["tmp_magma"] = hf_row["T_magma"]
+    # Update compositions
+    vol_dict = ConstructVolDict(hf_row, COUPLER_options)
+    for g in vol_dict.keys():
+        atmos.gas_vmr[g][:] = vol_dict[g]
+        atmos.gas_ovmr[g][:] = vol_dict[g]
 
-    # Solution type ~ surface state
-    cfg_toml["execution"]["solution_type"] = surf_state
+    # Update pressure grid 
+    atmos.p_boa = 1.0e5 * hf_row["P_surf"]
+    jl.AGNI.atmosphere.generate_pgrid_b(atmos)
 
-    # Tighter tolerances during first iters, to ensure consistent coupling
-    if loops_total < 3:
-        cfg_toml["execution"]["converge_rtol"] = 1.0e-3
-        
-    # Set plots 
-    cfg_toml["plots"]["at_runtime"]     = agni_debug and make_plots
-    cfg_toml["plots"]["temperature"]    = make_plots
-    cfg_toml["plots"]["fluxes"]         = make_plots
+    # Update surface temperature(s)
+    atmos.tmp_surf  = hf_row["T_surf"]
+    atmos.tmp_magma = hf_row["T_magma"]
 
-    # AGNI log level
-    cfg_toml["execution"]["verbosity"] = 1
-    # if agni_debug:
-    #     cfg_toml["execution"]["verbosity"] = 2
+    # Temperature profile 
+    if not resume:
+        jl.AGNI.setpt.isothermal_b(atmos, atmos.tmp_surf)
 
-    # Write new configuration file 
-    with open(cfg_this, 'w') as hdl:
-        toml.dump(cfg_toml, hdl)
+    return atmos
 
-    # ---------------------------
-    # Run AGNI
-    # ---------------------------
 
-    # Setup output stream
-    if agni_debug:
-        agni_stdout = sys.stdout 
-    else:
-        log.info("AGNI output suppressed (see agni.log)")
-        agni_stdout = subprocess.DEVNULL
-
-    # Call the module
-    log.debug("Call AGNI subprocess - output below...")
-    call_sequence = [ os.path.join(dirs["agni"],"agni.jl"), cfg_this]
-    proc = subprocess.run(call_sequence, stdout=agni_stdout, stderr=sys.stdout) 
-    
-    # Copy AGNI log into PROTEUS log. There are probably better ways to do this, but it 
-    #     works well enough. We don't use agni_debug much anyway
-    if agni_debug:
-        logpath = GetLogfilePath(dirs["output"], GetCurrentLogfileIndex(dirs["output"]))
-        with open(logpath, "a") as outfile:
-            with open(os.path.join(dirs["output"], "agni.log"), "r") as infile:
-                outfile.write(infile.read())
-    
-    return bool(proc.returncode == 0)
-
-def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
+def RunAGNI(atmos, loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
     """Run AGNI atmosphere model.
     
     Calculates the temperature structure of the atmosphere and the fluxes, etc.
@@ -212,6 +198,8 @@ def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
 
     Parameters
     ----------
+        atmos : atmosphere.Atmos_t
+            Atmosphere struct
         loops_total : int 
             Model total loops counter.
         dirs : dict
@@ -223,10 +211,29 @@ def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
 
     Returns
     ----------
+        atmos : atmosphere.Atmos_t
+            Atmosphere struct
         output : dict
             Output variables, as a dictionary
 
     """
+
+    # Chemistry 
+    chem_type = COUPLER_options["atmosphere_chemistry"]
+    if chem_type > 0:
+        if chem_type == 1:
+            # equilibrium
+            include_all = True
+
+        elif chem_type >= 2:
+            # kinetics 
+            raise Exception("Chemistry type %d unsupported by AGNI"%chem_type)
+    
+    # Solution type
+    surf_state = int(COUPLER_options["atmosphere_surf_state"])
+    if not (0 <= surf_state <= 3):
+        UpdateStatusfile(dirs, 20)
+        raise Exception("Invalid surface state %d" % surf_state)
 
     # Inform
     log.info("Running AGNI...")
@@ -255,9 +262,19 @@ def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
     while not agni_success:
         log.info("Attempt %d" % attempts)
 
-        # Try the module
-        agni_success = _try_agni(loops_total, dirs, COUPLER_options, hf_row, make_plots, 
-                                 easy_start, linesearch, dx_max, resume_prev)
+        # Try solving temperature profile
+        agni_success = jl.AGNI.solver.solve_energy_b(
+                            atmos, sol_type=surf_state,
+                            chem_type=chem_type, 
+                            conduct=False, convect=True, latent=True, sens_heat=True, 
+                            max_steps=200,
+                            max_runtime=600.0, 
+                            conv_atol=1e-3, conv_rtol=5e-3, 
+                            method=1, 
+                            dx_max=dx_max, ls_method=linesearch, easy_start=easy_start,
+                            save_frames=False
+                            )
+
 
         if agni_success:
             # success
@@ -277,26 +294,14 @@ def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
                 log.error("Maximum attempts when executing AGNI")
                 break
    
-    # Move files
-    log.debug("Tidy files")
-    files_move = [  
-                    ("atm.nc", "data/"+time_str+"_atm.nc"),
-                    ("agni.log", "agni_recent.log")
-                 ]
-    if make_plots:
-        files_move.append(("plot_fluxes.png", "plot_fluxes_atmosphere.png"))
-    for pair in files_move:
-        p_inp = os.path.join(dirs["output"], pair[0])
-        p_out = os.path.join(dirs["output"], pair[1])
-        safe_rm(p_out)
-        shutil.move(p_inp, p_out)
+    # Write output data 
+    ncdf_path = os.path.join(dirs["output"],"data",time_str+"_atm.nc")
+    jl.AGNI.dump.write_ncdf(atmos, ncdf_path)
 
-    # Remove files
-    files_remove = ["plot_ptprofile.png", "fl.csv", "ptz_ini.csv", "ptz.csv", "agni.toml", 
-                    "solver.png", "jacobian.png"] 
-    for frem in files_remove:
-        frem_path = os.path.join(dirs["output"],frem)
-        safe_rm(frem_path)
+    # Make plots 
+    if make_plots:
+        flux_path = os.path.join(dirs["output"],"plot_fluxes_atmosphere.png")
+        jl.AGNI.plotting.plot_fluxes(atmos, flux_path)
 
     # ---------------------------
     # Read results
@@ -304,13 +309,13 @@ def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
     
     log.debug("Read results")
     ds = nc.Dataset(os.path.join(dirs["output"],"data",time_str+"_atm.nc"))
-    net_flux =      np.array(ds.variables["fl_N"][:])
-    LW_flux_up =    np.array(ds.variables["fl_U_LW"][:])
-    SW_flux_up =    np.array(ds.variables["fl_U_SW"][:])
-    arr_p =         np.array(ds.variables["p"][:])
-    arr_z =         np.array(ds.variables["z"][:])
-    radius =        float(ds.variables["planet_radius"][:])
-    T_surf =        float(ds.variables["tmp_surf"][:])
+    net_flux =      np.array(atmos.flux_n)
+    LW_flux_up =    np.array(atmos.flux_u_lw)
+    SW_flux_up =    np.array(atmos.flux_u_sw)
+    arr_p =         np.array(atmos.p)
+    arr_z =         np.array(atmos.z)
+    radius =        float(atmos.rp)
+    T_surf =        float(atmos.tmp_surf)
     ds.close()
 
     # New flux from SOCRATES
@@ -337,5 +342,5 @@ def RunAGNI(loops_total:int, dirs:dict, COUPLER_options:dict, hf_row:dict):
     output["T_surf"] = T_surf
     output["z_obs"]  = z_obs + radius
     
-    return output
+    return atmos, output
